+++ src/App/components/Main/Main.js
@@ -1,13 +1,12 @@
 import React, { useState } from 'react';
 import { Header } from './Header/Header'
 import { Sections } from './Sections/Sections';
-import { SearchPanel } from './SearchPanels/SearchPanels';
+import { SearchPanel } from './SearchPanel/SearchPanel';
 
 function Main() {
     const [isMouseLeaved, setMouseLeaved] = useState(true);
     const [selectedTabName, setSelectedTabName] = useState('none');
 
-    // 날짜, 인원, 숙소타입, 가격, 즉시예약, 필터추가 탭(검색옵션탭)
     const setSelectedTab = (tabName) => {        
         setSelectedTabName(tabName);
     }
@@ -24,7 +23,6 @@ function Main() {
         if(isMouseLeaved === true && selectedTabName !== "none") { 
             setSelectedTabName("none"); 
         }
-        
     }
 
     return (
@@ -33,10 +31,7 @@ function Main() {
             <SearchPanel 
                 handleOnMouseLeave={handleOnMouseLeave} 
                 handleOnMouseEnter={handleOnMouseEnter} 
-
-                // 날짜, 인원, 숙소타입, 가격, 즉시예약, 필터추가
                 selectedTabName={selectedTabName} 
-
                 passSelectedTab={setSelectedTab}
             />
             <Sections />
