+++ src/App/components/Main/SearchPanel/SearchTabs/InnSearchTab.js
@@ -1,23 +1,26 @@
 import React from 'react';
 
 import { SearchTabStyle } from './SearchTabStyle'
-import { SearchOptionTabs } from './SearchOptionTabs/SearchOptionTabs'
-const { DateSetTab, GuestSetTab } = SearchOptionTabs;
+import { OptionTabButtons } from './SearchOptionTabs/OptionTabs/OptionTabButtons/OptionTabButtons';
+const { DateSetTab, GuestSetTab, InnTypeSetTab, PriceSetTab, InstantBookSetTab, FilterAddTab } = OptionTabButtons;
 
-
-function AllSearchTab(props) {
+function InnSearchTab(props) {
     const passButtonClick = event => props.passButtonClick(event);
     const passTabUrl = (optionTabUrl) => props.passTabUrl(optionTabUrl);
     const guestSetTabProps = { guestNum: props.guestNum, toddlerNum: props.toddlerNum }
     const passInfo = { match: props.match, passTabUrl: passTabUrl, passButtonClick: passButtonClick }
 
     return (
-        <SearchTabStyle >
-            <h4>AllSearchPanel</h4>
+        <SearchTabStyle>
+            <h4>InnSearchPanel</h4>
             <DateSetTab {...passInfo} />
             <GuestSetTab {...passInfo} {...guestSetTabProps} />
+            <InnTypeSetTab {...passInfo} />
+            <PriceSetTab {...passInfo} />
+            <InstantBookSetTab {...passInfo} />
+            <FilterAddTab {...passInfo} />
         </SearchTabStyle>
     )
 }
 
-export { AllSearchTab };
+export { InnSearchTab };
