+++ src/App/components/Main/SearchPanel/SearchTabs/SearchOptionTabs/OptionTabs/AddFilters.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { SearchOptionStyle } from './SearchOptionStyle';
+import { OptionTabStyle } from './OptionTabStyle';
 
 
 function AddFilters(props) {
@@ -9,9 +9,9 @@ function AddFilters(props) {
     const handleOnMouseEnter = () => props.handleOnMouseEnter();
 
     return (
-        <SearchOptionStyle onMouseLeave={handleOnMouseLeave} onMouseEnter={handleOnMouseEnter}>
+        <OptionTabStyle onMouseLeave={handleOnMouseLeave} onMouseEnter={handleOnMouseEnter}>
             필터를 추가하는 옵션 패널
-        </SearchOptionStyle>
+        </OptionTabStyle>
     )
 }
