+++ src/App/components/Main/SearchPanel/SearchTabs/SearchOptionTabs/OptionTabs/Calendar.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { SearchOptionStyle } from './SearchOptionStyle';
+import { OptionTabStyle } from './OptionTabStyle';
 
 function Calendar(props) {
 
@@ -7,9 +7,9 @@ function Calendar(props) {
     const handleOnMouseEnter = () => props.handleOnMouseEnter();
     
     return (
-        <SearchOptionStyle onMouseLeave={handleOnMouseLeave} onMouseEnter={handleOnMouseEnter}>
+        <OptionTabStyle onMouseLeave={handleOnMouseLeave} onMouseEnter={handleOnMouseEnter}>
             달력, 기간을 설정하는 옵션 패널
-        </SearchOptionStyle>
+        </OptionTabStyle>
     )
 }
