+++ src/App/components/Main/SearchPanel/SearchTabs/SearchOptionTabs/OptionTabs/InnType.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { SearchOptionStyle } from './SearchOptionStyle';
+import { OptionTabStyle } from './OptionTabStyle';
 
 
 function InnType(props) {
@@ -9,9 +9,9 @@ function InnType(props) {
     const handleOnMouseEnter = () => props.handleOnMouseEnter();
 
     return (
-        <SearchOptionStyle onMouseLeave={handleOnMouseLeave} onMouseEnter={handleOnMouseEnter}>
+        <OptionTabStyle onMouseLeave={handleOnMouseLeave} onMouseEnter={handleOnMouseEnter}>
             숙소 타입을 설정하는 옵션 패널
-        </SearchOptionStyle>
+        </OptionTabStyle>
     )
 }
