+++ src/App/components/Main/SearchPanel/SearchTabs/SearchOptionTabs/OptionTabs/InstantBook.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { SearchOptionStyle } from './SearchOptionStyle';
+import { OptionTabStyle } from './OptionTabStyle';
 
 function InstantBook(props) {
 
@@ -8,9 +8,9 @@ function InstantBook(props) {
     const handleOnMouseEnter = () => props.handleOnMouseEnter();
 
     return (
-        <SearchOptionStyle onMouseLeave={handleOnMouseLeave} onMouseEnter={handleOnMouseEnter}>
+        <OptionTabStyle onMouseLeave={handleOnMouseLeave} onMouseEnter={handleOnMouseEnter}>
             즉시예약을 설정하는 옵션 패널
-        </SearchOptionStyle>
+        </OptionTabStyle>
     )
 }
