+++ src/App/components/Main/SearchPanel/SearchTabs/SearchTabStyle.js
